Many changes - mostly reducing redundancy in shooter, collector, and related
Renamed CollectBall to IntakeBall. Created CollectorMove function with
PID control. Condensed ShootHighGoal and ShootLowGoal into ShootGoal
command. Created command groups to intake the ball, shoot the high goal
and shoot the low goal (need values for the collector arm encoder to aim
for the last two). Messed around with PortcullisAuto and broke it. THIS
NEEDS A COMMAND GROUP MADE FOR IT.  Created PIDWrite and PIDGet
functions in Collector. Created ActivateKicker, ActivateShooter,
ActivateCollector, and getRollerSpeed in Collector (these still need
code).